Title: 두 실수의 곱, Time: 99ms, Memory: 9MB, Status: Wrong Answer - Codetree

+++ 250529/두 실수의 곱/the-product-of-two-real-numbers.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
         // Please write your code here.
-        int a = 5.26;
-        int b = 8.27;
+        double a = 5.26;
+        double b = 8.27;
 
         System.out.printf("%.2f", a*b);
     }
